3.22晚->支付模块完成

+++ CharityPlatform/src/main/java/com/seu/beyondtheboundary/charityplatform/domain/Project.java
@@ -25,10 +25,10 @@ public class Project implements Serializable {
     private Long status = 1L;  //状态,状态1表示待审核，2表示已审核通过未发布，3，表示审核通过已发布，0表示已审核未通过
 
     @Column(name ="aimDonation")
-    private Long aimDonation = 1L;  //目标金额
+    private float aimDonation = 1L;  //目标金额
 
     @Column(name ="alreadyDonation")
-    private Long alreadyDonation = 1L;  //目前已筹
+    private float alreadyDonation = 1L;  //目前已筹
 
     @Column(name = "reading")
     private Long reading = 0L;  //阅读量
@@ -119,19 +119,19 @@ public class Project implements Serializable {
         this.status = status;
     }
 
-    public Long getAimDonation() {
+    public float getAimDonation() {
         return aimDonation;
     }
 
-    public void setAimDonation(Long aimDonation) {
+    public void setAimDonation(float aimDonation) {
         this.aimDonation = aimDonation;
     }
 
-    public Long getAlreadyDonation() {
+    public float getAlreadyDonation() {
         return alreadyDonation;
     }
 
-    public void setAlreadyDonation(Long alreadyDonation) {
+    public void setAlreadyDonation(float alreadyDonation) {
         this.alreadyDonation = alreadyDonation;
     }
